scripts/training/requirements.txt
# Edge Impulse Training Automation
# Install: pip install -r requirements.txt

# Edge Impulse SDK
edgeimpulse

# Google Cloud TTS (for synthetic voice generation)
google-cloud-texttospeech

# Audio processing
pydub

# HTTP requests
requests
httpx

# Environment variables
python-dotenv
